requirements.txt
numpy==1.23.5
scikit-learn
pandas
joblib
djangorestframework
pysafebrowsing